Moved common test fixtures to header file
